requirements.txt
earthengine-api==0.1.384
streamlit==1.29.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
opencv-python==4.8.1.78
rasterio==1.3.9
geopandas==0.14.1
folium==0.15.1
streamlit-folium==0.15.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-cloud-storage==2.10.0 